requirements.txt
appnope==0.1.0
boto==2.42.0
Collectfast==0.2.3
decorator==4.0.9
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.5
django-appconf==1.0.2
django-bootstrap3==7.0.1
django-localflavor==1.2
django-photologue==3.5.1
django-redis-cache==1.6.5
django-sortedm2m==1.3.2
django-storages==1.5.0
django-storages-redux==1.3
django-taggit==0.20.2
django-toolbelt==0.0.1
ExifRead==2.1.2
gnureadline==6.3.3
gunicorn==19.4.5
ipython==4.1.2
ipython-genutils==0.1.0
pexpect==4.0.1
pilkit==1.1.13
Pillow==3.3.0
psycopg2==2.6.1
python-dateutil==2.5.3
pytz==2016.6.1
redis==2.10.5
six==1.10.0
static3==0.7.0
